Factor out logging part from Network module
